# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SchemaId": {
				"type": "string",
				"example": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0"
			},
			"Version": {
				"type": "string",
				"example": "1.0.0"
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@aries-framework/rest",
		"version": "0.9.4",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/polygon/create-keys": {
			"post": {
				"operationId": "CreateKeyPair",
				"responses": {
					"200": {
						"description": "Secp256k1KeyPair",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"type": "string"
										},
										"publicKeyBase58": {
											"type": "string"
										},
										"privateKey": {
											"type": "string"
										}
									},
									"required": [
										"address",
										"publicKeyBase58",
										"privateKey"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create Secp256k1 key pair for polygon DID",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/polygon/create-schema": {
			"post": {
				"operationId": "CreateSchema",
				"responses": {
					"200": {
						"description": "Schema JSON",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create polygon based W3C schema",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"schema": {
										"additionalProperties": false,
										"type": "object"
									},
									"schemaName": {
										"type": "string"
									},
									"did": {
										"type": "string"
									}
								},
								"required": [
									"schema",
									"schemaName",
									"did"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/polygon/estimate-transaction": {
			"post": {
				"operationId": "EstimateTransaction",
				"responses": {
					"200": {
						"description": "Transaction Object",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Estimate transaction",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"transaction": {},
									"operation": {}
								},
								"required": [
									"transaction",
									"operation"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/polygon/{did}/{schemaId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "Schema Object",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Fetch schema details",
				"tags": [
					"Polygon"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/schemas/{schemaId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "Schema",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"ver": "1.0",
											"id": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0",
											"name": "schema",
											"version": "1.0",
											"attrNames": [
												"string"
											],
											"seqNo": 351936
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieve schema by schema id",
				"tags": [
					"Schemas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SchemaId"
						}
					}
				]
			}
		},
		"/schemas": {
			"post": {
				"operationId": "CreateSchema",
				"responses": {
					"200": {
						"description": "schema",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"ver": "1.0",
											"id": "WgWxqztrNooG92RXvxSTWv:2:schema_name:1.0",
											"name": "schema",
											"version": "1.0",
											"attrNames": [
												"string"
											],
											"seqNo": 351936
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Creates a new schema and registers schema on ledger",
				"tags": [
					"Schemas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"endorserDid": {
										"type": "string"
									},
									"endorse": {
										"type": "boolean"
									},
									"attributes": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"version": {
										"$ref": "#/components/schemas/Version"
									},
									"name": {
										"type": "string"
									},
									"issuerId": {
										"type": "string"
									}
								},
								"required": [
									"attributes",
									"version",
									"name",
									"issuerId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}